Remove JavaScript completely
That also includes the code to prepend "[LOCAL] " to the title of the website home page if it's hosted on `localhost` or `127.0.0.1`.

Now my website fully works without JavaScript (for now.) I also removed the warning that shows up when JavaScript is missing, since it's not needed at the moment.